Simplify shuffling code
Note: It might be better to have 'shuffle' actually create a shuffled
playlist, go through all the files in that order, and have it be
reshuffled automatically every time it has finished playing. That would
ensure that all files are played an equal number of times. In that case,
we could have shuffle and repeat be two distinct parameters.